Add files via upload
The main.py file contains the main loop of the program, the construction of the object. The settings.py file creates several variables and sets the screen/window of the program. The constants.py file contains some constants of the program. The typing.py file contains the 'typing' class with all its methods.